+++ example/lib/main.dart
@@ -1,7 +1,9 @@
 import 'package:example/pages/demo_page.dart';
 import 'package:flutter/material.dart';
+import 'package:flutter_web_plugins/url_strategy.dart';
 
 void main() {
+  usePathUrlStrategy();
   runApp(const MyApp());
 }