+++ lib/flutter_context_menu.dart
@@ -6,5 +6,6 @@ export 'src/core/models/context_menu.dart';
 export 'src/core/models/context_menu_entry.dart';
 export 'src/core/models/context_menu_item.dart';
 export 'src/core/utils/helpers.dart';
+export 'src/core/utils/menu_route_options.dart';
 export 'src/widgets/context_menu_region.dart';
 export 'src/widgets/context_menu_state.dart';