+++ lib/src/core/utils/helpers.dart
@@ -3,23 +3,18 @@ import 'package:flutter/widgets.dart';
 import '../../widgets/context_menu_state.dart';
 import '../../widgets/context_menu_widget.dart';
 import '../models/context_menu.dart';
+import 'menu_route_options.dart';
 
 /// Shows the root context menu popup.
 Future<T?> showContextMenu<T>(
   BuildContext context, {
   required ContextMenu<T> contextMenu,
-  RouteSettings? routeSettings,
-  bool? opaque,
-  bool? barrierDismissible,
-  Color? barrierColor,
-  String? barrierLabel,
-  Duration? transitionDuration,
-  Duration? reverseTransitionDuration,
-  RouteTransitionsBuilder? transitionsBuilder,
-  bool allowSnapshotting = true,
-  bool maintainState = false,
+  MenuRouteOptions? routeOptions,
 }) async {
   final menuState = ContextMenuState(menu: contextMenu);
+  routeOptions ??= const MenuRouteOptions(
+    barrierDismissible: true,
+  );
   return await Navigator.push<T>(
     context,
     PageRouteBuilder<T>(
@@ -28,22 +23,18 @@ Future<T?> showContextMenu<T>(
           children: [ContextMenuWidget(menuState: menuState)],
         );
       },
-      settings: routeSettings ?? const RouteSettings(name: "context-menu"),
       fullscreenDialog: true,
-      barrierDismissible: barrierDismissible ?? true,
-      opaque: opaque ?? false,
-      transitionDuration: transitionDuration ?? Duration.zero,
-      reverseTransitionDuration: reverseTransitionDuration ?? Duration.zero,
-      barrierColor: barrierColor,
-      barrierLabel: barrierLabel,
-      transitionsBuilder: transitionsBuilder ?? _defaultTransitionsBuilder,
-      allowSnapshotting: allowSnapshotting,
-      maintainState: maintainState,
+      settings: routeOptions.routeSettings,
+      barrierDismissible: routeOptions.barrierDismissible,
+      opaque: routeOptions.opaque,
+      transitionDuration: routeOptions.transitionDuration,
+      reverseTransitionDuration: routeOptions.reverseTransitionDuration,
+      barrierColor: routeOptions.barrierColor,
+      barrierLabel: routeOptions.barrierLabel,
+      transitionsBuilder: routeOptions.transitionsBuilder,
+      allowSnapshotting: routeOptions.allowSnapshotting,
+      maintainState: routeOptions.maintainState,
+      requestFocus: routeOptions.requestFocus,
     ),
   );
-}
-
-Widget _defaultTransitionsBuilder(
-    context, animation, secondaryAnimation, child) {
-  return child;
 }